// SDKs/googleapis.com/managedidentities/v1alpha1/typescript/src/sdk/models/operations/managedidentitiesprojectslocationsglobalpeeringscreate.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import * as shared from "../shared";


export class ManagedidentitiesProjectsLocationsGlobalPeeringsCreatePathParams extends SpeakeasyBase {
  @Metadata({ data: "pathParam, style=simple;explode=false;name=parent" })
  parent: string;
}


export class ManagedidentitiesProjectsLocationsGlobalPeeringsCreateQueryParams extends SpeakeasyBase {
  @Metadata({ data: "queryParam, style=form;explode=true;name=$.xgafv" })
  dollarXgafv?: shared.XgafvEnum;

  @Metadata({ data: "queryParam, style=form;explode=true;name=access_token" })
  accessToken?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=alt" })
  alt?: shared.AltEnum;

  @Metadata({ data: "queryParam, style=form;explode=true;name=callback" })
  callback?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=fields" })
  fields?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=key" })
  key?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=oauth_token" })
  oauthToken?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=peeringId" })
  peeringId?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=prettyPrint" })
  prettyPrint?: boolean;

  @Metadata({ data: "queryParam, style=form;explode=true;name=quotaUser" })
  quotaUser?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=uploadType" })
  uploadType?: string;

  @Metadata({ data: "queryParam, style=form;explode=true;name=upload_protocol" })
  uploadProtocol?: string;
}


export class ManagedidentitiesProjectsLocationsGlobalPeeringsCreateSecurity extends SpeakeasyBase {
  @Metadata({ data: "security, scheme=true;type=oauth2" })
  oauth2: shared.SchemeOauth2;

  @Metadata({ data: "security, scheme=true;type=oauth2" })
  oauth2c: shared.SchemeOauth2c;
}


export class ManagedidentitiesProjectsLocationsGlobalPeeringsCreateRequest extends SpeakeasyBase {
  @Metadata()
  pathParams: ManagedidentitiesProjectsLocationsGlobalPeeringsCreatePathParams;

  @Metadata()
  queryParams: ManagedidentitiesProjectsLocationsGlobalPeeringsCreateQueryParams;

  @Metadata({ data: "request, media_type=application/json" })
  request?: shared.Peering;

  @Metadata()
  security: ManagedidentitiesProjectsLocationsGlobalPeeringsCreateSecurity;
}


export class ManagedidentitiesProjectsLocationsGlobalPeeringsCreateResponse extends SpeakeasyBase {
  @Metadata()
  contentType: string;

  @Metadata()
  operation?: shared.Operation;

  @Metadata()
  statusCode: number;
}
